Sutori.AST.Utils

Description

Documentation

expressionType :: SutExpression -> SutType

Extracts the type from any expression

withPrimitiveType :: SutPrimitive -> SutExpression -> SutExpression

Clones the expression overriding the type for the given primitive type This is used for type coersion, mostly

asTypeError :: SutExpression -> SutExpression

Clones the expression but sets the type as a TypeError