Sutori.Error

Description

Documentation

data SutError

Different possible Sutori Errors

Constructors

LexicalError
An unknown or malformed token was read

GrammaticalError
The source code does not follow Sutori grammar

TypeError
A Type error occurred

UndefinedSymbolError
A symbol was used before it was defined

ArgumentsNumberError
A function was called with the wrong number of arguments

DuplicateSymbolError
A symbol was defined twice in the same scope

InternalError
The compiler failed for a known reason (not the user's fault)

NoError
The lack of error

logError :: SutLogger -> [(SutError, SutLog)]

The `error' log

lexerError :: String -> SutMonad ()

Logs a lexical error and continues

parserError :: SutToken -> SutMonad a

Logs a parsing error and throws out

typeError :: SutExpression -> SutType -> SutType -> String -> SutMonad ()

Logs a type error and continues

undefinedError :: SutID -> SymbolCat -> String -> SutMonad ()

Logs an undefined symbol error and continues

duplicateSymbolError :: SutSymbol a => a -> SutMonad ()

Logs a duplicate symbol error and continues

duplicateMemberError :: SutID -> SutMonad ()

Logs a duplicate member error and continues

argumentsNumberError :: SutID -> Int -> Int -> SutMonad ()

Logs a wrong number of parameters error and continues