Sutori.Lexer

Description

Documentation

lexerScanClean :: SutMonad SutToken

Scans for next token, but passes it though checks first

runLexer :: Options -> String -> SutMonad a -> Either (SutError, SutLog) (a, SutLogger)

Run the lexer on a given input string, with a given function

runLexerScan :: Options -> String -> Either (SutError, SutLog) ([SutToken], SutLogger)

Run the lexer on a given string, get the results

lexwrap :: (SutToken -> SutMonad a) -> SutMonad a

Run the lexer, but receive a continuation (Used by Happy)