Sutori.Lexer.Internals

Description

Documentation

utf8Encode :: Char -> [Word8]

Encode a Haskell String to a list of Word8 values, in UTF8 format. Taken from Alex wrapper templates

type SutoriInput = (SutPosn, Char, [Word8], String)

The type of the scanner's food (Current position, Previous character, pending bytes on current character, current input)

type AlexInput = SutoriInput

Alias needed by Alex

alexInputPrevChar :: SutoriInput -> Char

Needed by alex for look-back. We don't use look-back now

alexGetByte :: SutoriInput -> Maybe (Word8, SutoriInput)

Needed by alex to read

ignorePendingBytes :: SutoriInput -> SutoriInput

Ignore the current character's pending bytes

sutTabSize :: Int

Hard-coded tab size for positioning

lexerMove :: SutPosn -> Char -> SutPosn

Move the scanner position given a character

lexerGetInput :: SutMonad SutoriInput

Read input

lexerSetInput :: SutoriInput -> SutMonad ()

Set remaining input