Sutori.Logger

Description

Provides a common framework (class) to showSut sutori componentselementsconstructs.

Documentation

data SutLog

Simple tree data structure to allow pretty printing of sutori logs

Constructors

SutLogLeave String
A final message

SutLogNode String [SutLog]
A message with children messages

class SutShow a where

Interface for pretty-printing Sutori constructs (token, actions, tables, ...)

Minimal complete definition

showSut

Methods

showSut :: a -> SutLog

Allows to pretty-print a Sutori (internal or external) element

data SutLogger

Sutori logger, where we register logs for reasons, via a WriterT

Constructors

SutLogger [SutLog] [SutLog] [(SutError, SutLog)]

fromLeave :: SutLog -> String

Extract log strings