Sutori.Monad

Description

Documentation

type SutMonad a = ReaderT Options (StateT SutState (WriterT SutLogger (Except (SutError, SutLog)))) a

The Sutori monad. Composes state, logging and exception handling

data SutState

Monadic Lexer/Parser current state.

Constructors

SutState
Current error code, if any

Fields

lexerPosn :: SutPosn

Position at current input location

lexerInput :: String

Current input

lexerChar :: Char

Character before the input

lexerStateCode :: Int

Current startcode

lexerBytes :: [Word8]

Current bytes read

lexerDepth :: Int

Current comment nesting

lexerString :: String

Currently built string

lexerStringOn :: Bool

True if currently on an open strin

parserTable :: SymTable

The symtable

parserStack :: [Scope]

The scopes stack

parserScopes :: Set Scope

The set of open scopes

parserOffset :: Offset

Current offset to set on new local mem allocations

parserOffsetStk :: [Offset]

Stack of past scope offsets (to return to, once the scopes are closed

parserNextScope :: Scope

The next scope ID to open

mainModule :: SutModule

The main module (where compilation began)

tacTable :: TACTable

The table of generated TAC

tacNext :: Int

The next triplet index/The size of the current table

tacLabel :: Int

The next label for TAC generation

typesGraph :: TypeGraph

The constructed type graph

typesNextID :: SutTypeID

The next type ID to be introduced

logVerbose :: Bool

Set the output to verbose of not

errorCode :: SutError

runSutMonad :: SutMonad a -> Options -> SutState -> Except (SutError, SutLog) ((a, SutState), SutLogger)

Run the monad with a given action

parserCurrentScope :: SutState -> Scope

Get the current open scope

setErrorCode :: SutError -> SutMonad ()

Set the current error code for the compilation

initialSutoriState :: SutState

Initial state of a Sutori parsescanrun

insertScope :: SutMonad ()

Inserts a new scope into the parse

Inserts the scope into the open scopes set.

Pushes the scope into the open scopes stack.

Saves the current offset to return to it once the scope is closed.

removeScope :: SutMonad ()

Removes last scope from the parse

Pops closing scope from stack of open scopes.

Removes closing scope from set of open scopes.

Recovers previous offset to write new data over just-closed scope

whenVerbose :: SutMonad () -> SutMonad ()

Run a monadic action only if verbose if turned on

functionStarts :: SutMonad ()