Sutori.Parser.Expressions

Description

Documentation

type ExprTransform = SutExpression -> SutExpression

Represents a transformation from an expression to another

Ex. Change of type to TypeError

type SutUnaryOp = SutExpression -> SutMonad SutExpression

Represents a unary operation expression constructor

type SutBinaryOp = SutExpression -> SutExpression -> SutMonad SutExpression

Represents a binary operation expression constructor

Expressions

Literals

exprLiteral :: SutPrimitive -> SutLiteral -> SutMonad SutExpression

Generates AST expression for the given literal

literalBool :: Bool -> SutMonad SutExpression

literalChar :: String -> SutMonad SutExpression

literalInt :: Int -> SutMonad SutExpression

literalFloat :: Float -> SutMonad SutExpression

literalString :: String -> SutMonad SutExpression

constructStruct :: [(SutID, SutExpression)] -> SutMonad SutExpression

Constructs a struct from the list of ID -> Expression mappings

Right now, we implement these by storing the member IDs into the type graph

Note: We do not allow duplicated member IDs

memberType :: (SutID, SutType) -> SutMonad (SutID, SutTypeID)

Transforms a member with its type definition to a member with its type ID

Simple Operations

Unary operations

unaryPlus :: SutUnaryOp

unaryMinus :: SutUnaryOp

unaryNot :: SutUnaryOp

Helpers

generalizeExprType :: SutExpression -> SutExpression -> SutType

Gives the most general type from two expressions to which they can be casted

binaryOp

Arguments

:: ExprTransform
Transform for first operand

-> ExprTransform
Transform for second operand

-> ExprTransform
Transform result

-> SutOperator
Operator

-> SutBinaryOp
Both operands and result

General constructor for binary operation with the given operator applying the given checks

numericBinaryOp :: SutOperator -> SutBinaryOp

Binary operation specific constructor check

booleanBinaryOp :: SutOperator -> SutBinaryOp

Binary operation specific constructor check

sortBinaryOp :: SutOperator -> SutBinaryOp

Binary operation specific constructor check

eqBinaryOp :: SutOperator -> SutBinaryOp

Binary operation specific constructor check

Numerical binary operations

opAddition :: SutBinaryOp

opSubstraction :: SutBinaryOp

opMultiplication :: SutBinaryOp

opDivision :: SutBinaryOp

opIntDivision :: SutBinaryOp

opModulo :: SutBinaryOp

opPower :: SutBinaryOp

Sorting/Comparison operations

opGreaterEqual :: SutBinaryOp

(>=) receives two "sortable" types

opLessEqual :: SutBinaryOp

(<=) receives two "sortable" types

opLess :: SutBinaryOp

(<) receives two "sortable" types

opGreater :: SutBinaryOp

(>) receives two "sortable" types

Boolean binary operations

opAnd :: SutBinaryOp

Boolean AND receives two booleans

opOr :: SutBinaryOp

Boolean OR receives two booleans

opEqual :: SutBinaryOp

Equality check receives two "equalable" types

opNotEqual :: SutBinaryOp

NotEquality check receives two "equalable" types

Complex operations

assignment :: SutBinaryOp

An assignement to the left side of the value from the right side

This checks that a general type for converting the right side exists

Note: An assignment is both an instruction and an expression

Note: It is known the left side is assignable

Note: Right side must be of a more specific type than left side

arrayGet :: SutBinaryOp

The indexation of the array in the left side with the key/index on the right side

This checks the left side is actually of an array type

Note: It is known the left side is assignable and the right side is int expression

Note: Left side must be array type

memberGet :: SutExpression -> SutID -> SutMonad SutExpression

The access to a membercomponent (ID right) of a thingmachine (left)

This checks left side is either a Machine or a Thing, and that it knows a member with the ID

Note: It is known the left lise is assignable

createPointer :: SutID -> SutTypeID -> SutMonad SutExpression

Creates a direction to a value of type given

Note: We are already given the existent type ID

Note: Left side is known to be a person

dereference :: SutUnaryOp

Creates a dereferencing expression for the given expression

This checks the expression is a direction to somewhere

Note: It is know the expression is assignable

functionCall :: SutID -> [SutExpression] -> SutMonad SutExpression

A call to a function (ID) with parameters (given list)

This checks for the correct number of arguments and their types

Note: Left side is known to be existent function (?)

checkParamTypes :: [SutType] -> [SutExpression] -> SutMonad ()

Compares the given formal and actual parameters types

constructArray :: [SutExpression] -> SutMonad SutExpression

Constructs an array from the list of expressions All expressions must be of the same type

Note: Sutori grammar doesn't allow empty arrays

checkArrayType :: SutExpression -> SutType -> [SutType] -> SutMonad ()

Checks the type of the array is consistent with the types of its elements