Sutori.Parser.Symbols

Description

Documentation

findTypeID :: SutType -> SutMonad SutTypeID

Finds the existent typeID or inserts the type and gets the new ID

findType :: SutID -> SutMonad SutTypeID

Finds the typeID from a SutID

findExistentType :: SutTypeID -> SutMonad SutType

Finds an existent type from its ID

findPerson :: SutID -> SutMonad SutID

Checks if the given person already exists

findFunction :: SutID -> SutMonad SutID

Checks if the given function already exists

findVariable :: SutID -> SutMonad SutExpression

Checks if the given variable already exists, returns the ID wrapped as an expression