Sutori.Parser.TypeCheck

Description

Documentation

checkPrimitiveType :: (SutPrimitive -> SutPrimitive) -> SutExpression -> SutExpression

Takes a conversion function from primitive to primitive and checks/converts the given expression

checkNumeric :: SutExpression -> SutExpression

Checks if expression is or can be converted to numeric

checkEq :: SutExpression -> SutExpression

Checks if expression is checked for equality

checkSortable :: SutExpression -> SutExpression

Checks if expression is or can be converted to a sortable type

checkBoolean :: SutExpression -> SutExpression

Checks if expression is or can be converted to boolean

checkIndex :: SutExpression -> SutExpression

Checks if expression is or can be converted to an index type

checkInt :: SutExpression -> SutExpression

Checks if expression is or can be converted to integer (index type)

checkFloat :: SutExpression -> SutExpression

Checks if expression is or can be converted to float

checkPrintable :: SutExpression -> SutExpression

Checks if expression is or can be converted to string for printing