Sutori.SymTable

Description

This module defines all the interaction with the table, along with the symbols themselves, their different categories and payloads.

Documentation

class SutSymbol a where

Represents a basic symbol

Minimal complete definition

symID, symCat

Methods

symID

Arguments

:: a

-> SutID
The symbol ID

symCat

Arguments

:: a

-> SymbolCat
The symbol category

symScope

Arguments

:: a

-> Scope
The scope the symbol is defined. Zero by default

data SutSymbol'

Boxed SutSymbols for the table to handle

Constructors

SymModule' SymModule

SymFunction' SymFunction

SymPerson' SymPerson

SymVariable' SymVariable

SymType' SymType

class SutSymbol a => ASTSymbol a where

Represents a symbol that holds an AST

Minimal complete definition

symAST

Methods

symAST

Arguments

:: a

-> SutAST
The AST held by the symbol

symPreAST

Arguments

:: a

-> SutAST
An optional (empty by default) AST (hidden, preppended)

class SutSymbol a => ParametricSymbol a where

Represents a symbol which also has some parameters

Minimal complete definition

symParams

Methods

symParams

Arguments

:: a

-> [SutParam]
The list of parameters (in order)

type Scope = Int

A Scope is just a number uniquely representing a scope (NOT a level, but an ID)

data SutParam

A SutParam represents a function parameter, and it has a kind, type and ID

Constructors

SutParam

Fields

isRef :: Bool

Whether it is a reference or a value parameter

paramType :: SutTypeID

The type of the parameter

paramID :: SutID

The ID of the parameter

data SymFunction

A (top-level) function

Constructors

SymFunction SutID SutTypeID [SutParam] SutAST (Maybe SutAST)

data SymModule

A (top-level) module

Constructors

SymModule SutID SutAST

data SymPerson

A person who may do things in the story

Constructors

SymPerson SutID Scope

type SymTable = Map SutID [SutSymbol']

A SymTable represents the SymTable of a Sutori parsing, so it matches a SutID (symbol name) to a list of symbols

data SymType

A reference to a (now named) type

Constructors

SymType SutID Scope SutTypeID

data SymVariable

A variable

Constructors

SymVariable SutID Scope SutTypeID

data SymbolCat

Constructors

CatModule

CatPerson

CatType

CatVariable

CatFunction

CatMember

class SutSymbol a => TypedSymbol a where

Represents a symbol with a type

Minimal complete definition

symType

Methods

symType

Arguments

:: a

-> SutTypeID
The type of the symbol

insertSymbol :: SutSymbol' -> SymTable -> SymTable

Inserts a symbol into its ID stack in the map

lookupSymbols :: SutID -> SymTable -> Maybe [SutSymbol']

Lookup the list of symbols using the given ID

lookupSymbolsVariable :: SutID -> SymTable -> [SymVariable]

Lookup the list of variables with an ID

lookupSymbolsPerson :: SutID -> SymTable -> [SymPerson]

Lookup the list of persons with an ID

lookupSymbolsModule :: SutID -> SymTable -> [SymModule]

Lookup the list of modules with an ID

lookupSymbolsType :: SutID -> SymTable -> [SymType]

Lookup the list of types with an ID

lookupSymbolsFunction :: SutID -> SymTable -> [SymFunction]

Lookup the list of functions with an ID

updateSymbol :: SutSymbol' -> SymTable -> SymTable

Updates (read: replaces) the head of the symbol stack (for this symbol's ID

paramByValue :: SutTypeID -> SutID -> SutParam

Helper to abstract the True/False isRef field

paramByRef :: SutTypeID -> SutID -> SutParam

Helper to abstract the True/False isRef field

lookupAllFunctions :: Scope -> SymTable -> [SymFunction]

Lookup all defined functions in a given scope