Sutori.SymTable.Actions

Description

Documentation

isFunction :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isMember :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isModule :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isType :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isPerson :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isVariable :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

insert :: SymTable -> Scope -> SutSymCategory -> SutTypeID -> SutSymOther -> [SutID] -> SymTable

Inserts new symbols into the table (from a list of IDs)

insertSymbol :: SymTable -> SutSymbol -> SymTable

Inserts symbol into the table the given constructed symbol

insertParams :: SymTable -> Scope -> SutSymCategory -> [SutParam] -> SymTable

Inserts new symbols into the table (from a list of params)

symTypeDef :: SutSymbol -> SutTypeID

Extracts a type symbol type definition (SutTypeID)

lookupID :: SymTable -> SutID -> [SutSymbol]

Search for all symbols that match a SutID