Sutori.SymTable.Logger

Description

Orphan instances

SutShow SutParamKind
A parameter kind can be printed nicely

Methods

showSut :: SutParamKind -> SutLog

SutShow SutParam
A parameter can be printed nicely

Methods

showSut :: SutParam -> SutLog

SutShow SutSymbol
A sutori symbol in the table can be printed nicely

Methods

showSut :: SutSymbol -> SutLog

SutShow SutSymOther
The extra payload of a symbol can be printed nicely

Methods

showSut :: SutSymOther -> SutLog

SutShow SutSymCategory
A symbol category can be printed nicely

Methods

showSut :: SutSymCategory -> SutLog