Sutori.SymTable.Symbol

Description

Documentation

class SutSymbol a where

Represents a basic symbol

Minimal complete definition

symID, symCat

Methods

symID

Arguments

:: a

-> SutID
The symbol ID

symCat

Arguments

:: a

-> SymbolCat
The symbol category

symScope

Arguments

:: a

-> Scope
The scope the symbol is defined. Zero by default

class SutSymbol a => ParametricSymbol a where

Represents a symbol which also has some parameters

Minimal complete definition

symParams

Methods

symParams

Arguments

:: a

-> [SutParam]
The list of parameters (in order)

class SutSymbol a => TypedSymbol a where

Represents a symbol with a type

Minimal complete definition

symType

Methods

symType

Arguments

:: a

-> SutTypeID
The type of the symbol

class SutSymbol a => ASTSymbol a where

Represents a symbol that holds an AST

Minimal complete definition

symAST

Methods

symAST

Arguments

:: a

-> SutAST
The AST held by the symbol

symPreAST

Arguments

:: a

-> SutAST
An optional (empty by default) AST (hidden, preppended)

class SutSymbol a => OffsetSymbol a where

Represents a symbol that _may_ hold an offset

Minimal complete definition

symOffset

Methods

symOffset :: a -> Offset

type Scope = Int

A Scope is just a number uniquely representing a scope (NOT a level, but an ID)

data SutParam

A SutParam represents a function parameter, and it has a kind, type and ID

Constructors

SutParam

Fields

isRef :: Bool

Whether it is a reference or a value parameter

paramType :: SutTypeID

The type of the parameter

paramID :: SutID

The ID of the parameter

data SymModule

A (top-level) module

Constructors

SymModule SutID SutAST

data SymFunction

A (top-level) function

Constructors

SymFunction SutID SutTypeID [SutParam] SutAST (Maybe SutAST)

data SymPerson

A person who may do things in the story

Constructors

SymPerson SutID Scope

data SymVariable

A variable

Constructors

SymVariable SutID Scope SutTypeID Offset

data SymType

A reference to a (now named) type

Constructors

SymType SutID Scope SutTypeID

data SutSymbol'

Boxed SutSymbols for the table to handle

Constructors

SymModule' SymModule

SymFunction' SymFunction

SymPerson' SymPerson

SymVariable' SymVariable

SymType' SymType

data SymbolCat

Constructors

CatModule

CatPerson

CatType

CatVariable

CatFunction

CatMember

isModule :: SymbolCat -> Bool

isFunction :: SymbolCat -> Bool

isPerson :: SymbolCat -> Bool

isType :: SymbolCat -> Bool

isVariable :: SymbolCat -> Bool

paramByValue :: SutTypeID -> SutID -> SutParam

Helper to abstract the True/False isRef field

paramByRef :: SutTypeID -> SutID -> SutParam

Helper to abstract the True/False isRef field

unBoxFunction :: SutSymbol' -> SymFunction

unBoxPerson :: SutSymbol' -> SymPerson

unBoxVariable :: SutSymbol' -> SymVariable

unBoxModule :: SutSymbol' -> SymModule

unBoxType :: SutSymbol' -> SymType