Sutori.SymTable.Table

Description

Documentation

type SymTable = Map SutID [SutSymbol']

A SymTable represents the SymTable of a Sutori parsing, so it matches a SutID (symbol name) to a list of symbols

insertSymbol :: SutSymbol' -> SymTable -> SymTable

Inserts a symbol into its ID stack in the map

updateSymbol :: SutSymbol' -> SymTable -> SymTable

Updates (read: replaces) the head of the symbol stack (for this symbol's ID

lookupSymbols :: SutID -> SymTable -> Maybe [SutSymbol']

Lookup the list of symbols using the given ID

lookupSymbolsVariable :: SutID -> SymTable -> [SymVariable]

Lookup the list of variables with an ID

lookupSymbolsFunction :: SutID -> SymTable -> [SymFunction]

Lookup the list of functions with an ID

lookupSymbolsModule :: SutID -> SymTable -> [SymModule]

Lookup the list of modules with an ID

lookupSymbolsType :: SutID -> SymTable -> [SymType]

Lookup the list of types with an ID

lookupSymbolsPerson :: SutID -> SymTable -> [SymPerson]

Lookup the list of persons with an ID

lookupAllFunctions :: Scope -> SymTable -> [SymFunction]

Lookup all defined functions in a given scope