Sutori.TAC

Description

Documentation

data TACTable

The actual generated code is a list of instruction triplets and a list of pointers to them

Constructors

TACTable
Table of triplets

Fields

tacInstructions :: [Int]

Table of pointers to triplets

tacTriplets :: [TAC]

data TACAddress

TACAddress for TAC Instructions

Constructors

TACGlobal SutID
An actual ID from the source code. Used for global (scope 0) variables

TACOffset Int
An offset from the current frame pointer

TACLit SutLiteral
A literal value either explicit in the source code or calculated on the run.

TACID Int
The TAC number that calculated the relevant expression.

TACLabel Int
An incremental label to some code.

TACFun SutID
A function name/label

data TACType

A TAC Instruction to be appended to the generated intermediate code

Constructors

Basic
For simple expressions.

Fields

tacOp :: SutOperator

SysCall
Perform a syscall (IO, Mem, else)

Fields

tacSys :: SutSys

Copy
Copy operation, from a address to another

Addr
Address for an indirection

Jump
Unconditional jump, just where to.

JumpUnless
Condition jump

Pointed
Assign an address from a indirect pointer.

Param
Stack a function parameter

Call
Call a function of so many (already stacked) parameters

Return
Return from a function (possibly a value).

data TAC

TAC Instruction

Constructors

TAC
An instruction

Fields

tacType :: TACType

Operation to perform

tac1 :: Maybe TACAddress

First address. Usually first operand.

tac2 :: Maybe TACAddress

Second address. Usually second operand.

Label Int
A numbered label to output literally

FunLabel SutID
A string label to output literally

addTAC :: TAC -> SutMonad TACAddress

Appends to the TAC table a new triplet, references it on the instructions TAC table

Returns the Address of the inserted TAC

genCode :: SutMonad TACTable

Generates code from the already built mainModule AST

Generates in order: TAC for main/global code, then TACs for each function.