Sutori.Types

Description

TODO: Sort and document API

Documentation

type SutTypeID = Int

A type will be represented by an ID

data SutType

Sutori type constructors

Constructors

SutPrimitiveType SutPrimitive
A Primitive Type

SutDirection SutTypeID
A Direction (Pointer) to a value of some type

SutChain Int SutTypeID
A Chain (Array) of fixed size of element type

SutMachine [(SutID, SutTypeID)]
A Machine (Struct) of different components

SutThing [(SutID, SutTypeID)]
A Thing (Union) that might have different names

primitiveType :: SutType -> SutPrimitive

Extract the primitive from the type definition

generalizeTypes :: SutType -> SutType -> SutType

Generalize two types to their LCA, if any

primitiveError :: SutType

Constant primitive error type

data TypeGraph

The type graph must be a biderectional map (as there's a bijection between type and its ID)

Constructors

TypeGraph (Map SutType (SutTypeID, Int)) (Map SutTypeID (SutType, Int))

initialTypeGraph :: TypeGraph

The initial value for the graph. That is, the primitives added

initialNextTypeID :: Int

The initial value for the "next type ID"

insertType :: (SutType, SutTypeID) -> TypeGraph -> TypeGraph

Inserts a type mapping into the map

lookupTypeID :: SutType -> TypeGraph -> Maybe (SutTypeID, Int)

Maybe the SutTypeID of the given type

lookupType :: SutTypeID -> TypeGraph -> Maybe (SutType, Int)

Maybe the type of the given SutTypeID

orderedGraph :: TypeGraph -> [(SutTypeID, (SutType, Int))]

The graph as a list sorted by ID (first appearance?)

data SutPrimitive

Sutori primitives

Constructors

SutBag
A bag (Int)

SutWallet
A wallet (Float)

SutPhrase
A phrase (String)

SutLight
A light (Bool)

SutLetter
A letter (Char)

SutTypeVoid
A void element (Void)

SutTypeError
A Type Error (&^%$#)

primitives :: [SutPrimitive]

Predefined Sutori types to initialize symtable

primitiveID :: SutPrimitive -> SutTypeID

Eventual TypeIDs for the type graph

primitiveIDs :: [(SutPrimitive, SutTypeID)]

Zipped primitives with their IDs

generalizePrimitives :: SutPrimitive -> SutPrimitive -> SutPrimitive

Given two types, this is the most general one (LCA) Right now LCA is just always one of them

toTypeLight :: SutPrimitive -> SutPrimitive

Convert primitive to boolean type

toTypeNum :: SutPrimitive -> SutPrimitive

Convert primitive to the most specific numerical type available

toTypeBag :: SutPrimitive -> SutPrimitive

Convert primitive to the most specific numerical type available

toTypeWallet :: SutPrimitive -> SutPrimitive

Convert primitive to the general float type

toTypePhrase :: SutPrimitive -> SutPrimitive

Convert primitive to the most general string (printable) type

toTypeSortable :: SutPrimitive -> SutPrimitive

Convert primitive to a sortable type or error

toTypeEq :: SutPrimitive -> SutPrimitive

Convert primitive to a equalable type or error

memberOffset :: SutType -> String -> Int

Lookup and check a member's offset