Sutori.Types.Primitives

Description

Documentation

data SutPrimitive

Sutori primitives

Constructors

SutBag
A bag (Int)

SutWallet
A wallet (Float)

SutPhrase
A phrase (String)

SutLight
A light (Bool)

SutLetter
A letter (Char)

SutTypeVoid
A void element (Void)

SutTypeError
A Type Error (&^%$#)

type SutTypeID = Int

A type will be represented by an ID

primitives :: [SutPrimitive]

Predefined Sutori types to initialize symtable

primitiveID :: SutPrimitive -> SutTypeID

Eventual TypeIDs for the type graph

primitiveIDs :: [(SutPrimitive, SutTypeID)]

Zipped primitives with their IDs

generalizePrimitives :: SutPrimitive -> SutPrimitive -> SutPrimitive

Given two types, this is the most general one (LCA) Right now LCA is just always one of them

toTypeLight :: SutPrimitive -> SutPrimitive

Convert primitive to boolean type

toTypeNum :: SutPrimitive -> SutPrimitive

Convert primitive to the most specific numerical type available

toTypeBag :: SutPrimitive -> SutPrimitive

Convert primitive to the most specific numerical type available

toTypeWallet :: SutPrimitive -> SutPrimitive

Convert primitive to the general float type

toTypePhrase :: SutPrimitive -> SutPrimitive

Convert primitive to the most general string (printable) type

toTypeSortable :: SutPrimitive -> SutPrimitive

Convert primitive to a sortable type or error

toTypeEq :: SutPrimitive -> SutPrimitive

Convert primitive to a equalable type or error