Sutori.Utils

Description

Documentation

anyf :: [a -> Bool] -> a -> Bool

Conjunction of a the results of a list of functions applied to a single value

allf :: [a -> Bool] -> a -> Bool

Conjunction of a the results of a list of functions applied to a single value

filterBy :: Ord a => ([a] -> Bool) -> [a] -> [[a]]

Filters a list with a given filter function

repeated :: Ord a => [a] -> [a]

Filters out unique elements in a list

repeatedBy :: (Eq c, Ord b) => (a -> c) -> (a -> b) -> [a] -> [a]

Filters consecutive uniques, given a grouping and a sorting function

type SutID = String

Global type for Sutori IDs