Sutori.Logger

Description

Provides a common framework (class) to showSut sutori componentselementsconstructs. Right now, error codes are defined here.

Documentation

data SutLog

Simple tree data structure to allow pretty printing of sutori logs

Constructors

SutLogLeave String
A final message

SutLogNode String [SutLog]
A message with children messages

class SutShow a where

Interface for pretty-printing Sutori constructs (token, actions, tables, ...)

Minimal complete definition

showSut

Methods

showSut :: a -> SutLog

Allows to pretty-print a Sutori (internal or external) element

data SutLogger

Sutori logger, where we register logs for reasons, via a WriterT

Constructors

SutLogger [SutLog] [SutLog] [(SutError, SutLog)]

data SutError

Different possible Sutori Errors

Constructors

LexicalError
An unknown or malformed token was read

GrammaticalError
The source code does not follow Sutori grammar

TypeError
A Type error occurred

UndefinedSymbolError
A symbol was used before it was defined

ArgumentsNumberError
A function was called with the wrong number of arguments

DuplicateSymbolError
A symbol was defined twice in the same scope

InternalError
The compiler failed for a known reason (not the user's fault)

NoError
The lack of error

fromLeave :: SutLog -> String

Extract log strings