Sutori.Parser.Definitions

Description

Documentation

defPerson :: SutID -> SutMonad ()

Includes a new person into the story

TODO: Refactor code. all these functions use mostly repeated code

defVariable :: SutID -> SutTypeID -> (SutID, Maybe SutExpression) -> SutMonad ()

Defines a new variable of given type and optionally assigns it an initial value.

TODO: Add assignment to AST (separate from definition, which does not go into AST)

defType :: SutID -> SutID -> SutTypeID -> SutMonad ()

Associates the SutID to the newly constructed type, assuming the name has not been used before

defModule :: SutID -> SutBlock -> SutMonad ()

Defines a module for importing/exporting

Right now it's basically a stub as there's only one module and the data structure only keeps the SutID and AST

defFunction :: SutTypeID -> SutBlock -> SutMonad ()

Defines a function with ID pushed and no parameters

defFunction' :: SutTypeID -> SutBlock -> SutMonad ()

Defines a function with ID and parameters already present

insertFunctionID :: SutID -> SutMonad SutID

Defines a new function with a given name (no parameters or body yet)

insertParam :: (SutParamKind, SutTypeID, SutID) -> SutMonad ()

Defines a new parameter in the last function defined