Sutori.SymTable

Description

This module defines all the interaction with the table, along with the symbols themselves, their different categories and payloads.

Documentation

data SutSymbol

A SutSymbol represents a symbol in the SymTable of a Sutori parsing. It is made of a Token, a Category, a Scope, a Type, and perhaps other information

Constructors

SutSymbol

Fields

symID :: SutID

The symbol ID

symCat :: SutSymCategory

The category of the symbol

symScope :: Scope

The scope the symbol is defined

symType :: SutTypeID

The type of the symbol, if any (might be void)

symOther :: SutSymOther

The extra payload some symbol kinds have

data SutSymCategory

A SutSymCategory. Can be shown by showSut

Constructors

CatModule
A module

CatFunction
A (top-level) function

CatPerson
A person who may do things in the story

CatVariable
A variable

CatParameter
A function parameter

CatType
A reference to a type

CatMember
A member of a structured type (does not go inot the SymTable)

data SutSymOther

A SutSymOther represents the extra payload of a symbol and it's either a function definition's AST, the parameter kind, the type definition or nothing, depending on the kind of symbol

Constructors

SymAST

Fields

otherASTParams :: [SutParam]

otherASTBlock :: SutBlock

SymParamKind

Fields

otherParamKind :: SutParamKind

SymTypeDef

Fields

otherTypeDef :: SutTypeID

SymNothing

type SymTable = Map SutID [SutSymbol]

A SymTable represents the SymTable of a Sutori parsing, so it matches a SutID (symbol name) to a list of symbols

type Scope = Int

A Scope is just a number uniquely representing a scope (NOT a level, but an ID)

data SutParamKind

A SutParamKind represents a function parameter kind, that is it can be passed either by reference or by value.

Constructors

SutRef

SutVal

data SutParam

A SutParam represents a function parameter, and it has a kind, type and ID

Constructors

SutParam

Fields

paramKind :: SutParamKind

Whether it is a reference or a value parameter

paramType :: SutTypeID

The type of the parameter

paramID :: SutID

The ID of the parameter

isFunction :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isMember :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isModule :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isType :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isPerson :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

isVariable :: SutSymCategory -> Bool

Boolean functions to check if a symbol if of some category

insert :: SymTable -> Scope -> SutSymCategory -> SutTypeID -> SutSymOther -> [SutID] -> SymTable

Inserts new symbols into the table (from a list of IDs)

insertSymbol :: SymTable -> SutSymbol -> SymTable

Inserts symbol into the table the given constructed symbol

insertParams :: SymTable -> Scope -> SutSymCategory -> [SutParam] -> SymTable

Inserts new symbols into the table (from a list of params)

symTypeDef :: SutSymbol -> SutTypeID

Extracts a type symbol type definition (SutTypeID)

lookupID :: SymTable -> SutID -> [SutSymbol]

Search for all symbols that match a SutID